remove nav selection modal
