+++ packages/builder/src/components/design/NavigationPanel/NavigationSelectionModal.svelte
@@ -1,98 +1,12 @@
 <script>
   import { ModalContent, Body, Detail } from "@budibase/bbui"
-  import { store, selectedAccessRole, allScreens } from "builderStore"
-  import { cloneDeep } from "lodash/fp"
-  import sanitizeUrl from "builderStore/store/screenTemplates/utils/sanitizeUrl"
-  import { onDestroy } from "svelte"
 
   export let selectedScreens
-  export let screenName
-  export let url
   export let chooseModal
-
-  let roleId = $selectedAccessRole || "BASIC"
-
-  let routeError
+  export let save
   let selectedNav
   let createdScreens = []
-  $: {
-    selectedScreens?.forEach(screen => {
-      createdScreens = [...createdScreens, screen.create()]
-    })
-  }
-
   $: blankSelected = selectedScreens.length === 1
-
-  const save = async screens => {
-    for (let screen of screens) {
-      await saveScreens(screen)
-    }
-
-    let navLayout = cloneDeep(
-      $store.layouts.find(layout => layout._id === "layout_private_master")
-    )
-    navLayout.props.navigation = selectedNav
-
-    await store.actions.routing.fetch()
-    await store.actions.layouts.save(navLayout)
-    selectedScreens = []
-    screenName = ""
-    url = ""
-  }
-  const saveScreens = async draftScreen => {
-    let existingScreenCount = $store.screens.filter(
-      s => s.props._instanceName == draftScreen.props._instanceName
-    ).length
-
-    if (existingScreenCount > 0) {
-      let oldUrlArr = draftScreen.routing.route.split("/")
-      oldUrlArr[1] = `${oldUrlArr[1]}-${existingScreenCount + 1}`
-      draftScreen.routing.route = oldUrlArr.join("/")
-    }
-
-    let route = url ? sanitizeUrl(`${url}`) : draftScreen.routing.route
-    if (draftScreen) {
-      if (!route) {
-        routeError = "URL is required"
-      } else {
-        if (routeExists(route, roleId)) {
-          routeError = "This URL is already taken for this access role"
-        } else {
-          routeError = ""
-        }
-      }
-
-      if (routeError) return false
-
-      if (screenName) {
-        draftScreen.props._instanceName = screenName
-      }
-
-      draftScreen.routing.route = route
-
-      await store.actions.screens.create(draftScreen)
-      if (draftScreen.props._instanceName.endsWith("List")) {
-        await store.actions.components.links.save(
-          draftScreen.routing.route,
-          draftScreen.routing.route.split("/")[1]
-        )
-      }
-    }
-  }
-
-  const routeExists = (route, roleId) => {
-    return $allScreens.some(
-      screen =>
-        screen.routing.route.toLowerCase() === route.toLowerCase() &&
-        screen.routing.roleId === roleId
-    )
-  }
-
-  onDestroy(() => {
-    selectedScreens = []
-    screenName = ""
-    url = ""
-  })
 </script>
 
 <ModalContent
